:bug: ça compile

+++ src/main/java/ulb/views/FoodPopupViewController.java
@@ -150,6 +150,6 @@ public class FoodPopupViewController implements ViewController {
 
 		void onBack();
 
-		int getServingQuantityValue(String food);
+		double getServingQuantityValue(String food);
 	}
 }
